Prevent CAT endpoint misconfiguration
If a BPF configuration CAT endpoint pair (IPv4 addresses and UDP port
numbers) is repeated, log an error and cause a LogF() to occur
in the BPF main program.

Add a unique 32-bit client ID field to the SLIQ connection handshake
header.  This is used to tell if connection endpoints are
misconfigured, with an endpoint (IPv4 address and UDP port number)
accidentally repeated.  When this happens, instead of the two
endpoints confusing one another, one endpoint will connect and the
other endpoint will get a connection reject each time it attempts to
connect.